LPS-68119 Move Polls tests
