+++ modules/apps/forms-and-workflow/polls/polls-test/src/testIntegration/java/com/liferay/polls/internal/exportimport/data/handler/test/PollsChoiceStagedModelDataHandlerTest.java
@@ -12,10 +12,12 @@
  * details.
  */
 
-package com.liferay.polls.lar.test;
+package com.liferay.polls.internal.exportimport.data.handler.test;
 
 import com.liferay.arquillian.extension.junit.bridge.junit.Arquillian;
+import com.liferay.polls.model.PollsChoice;
 import com.liferay.polls.model.PollsQuestion;
+import com.liferay.polls.service.PollsChoiceLocalServiceUtil;
 import com.liferay.polls.service.PollsQuestionLocalServiceUtil;
 import com.liferay.polls.util.test.PollsTestUtil;
 import com.liferay.portal.kernel.model.Group;
@@ -24,9 +26,11 @@ import com.liferay.portal.kernel.test.rule.AggregateTestRule;
 import com.liferay.portal.kernel.test.rule.Sync;
 import com.liferay.portal.kernel.test.rule.SynchronousDestinationTestRule;
 import com.liferay.portal.kernel.test.rule.TransactionalTestRule;
+import com.liferay.portal.kernel.util.LocaleUtil;
 import com.liferay.portal.lar.test.BaseStagedModelDataHandlerTestCase;
 import com.liferay.portal.test.rule.LiferayIntegrationTestRule;
 
+import java.util.HashMap;
 import java.util.List;
 import java.util.Map;
 
@@ -37,10 +41,11 @@ import org.junit.runner.RunWith;
 
 /**
  * @author Shinn Lok
+ * @author Mate Thurzo
  */
 @RunWith(Arquillian.class)
 @Sync
-public class PollsQuestionStagedModelDataHandlerTest
+public class PollsChoiceStagedModelDataHandlerTest
 	extends BaseStagedModelDataHandlerTestCase {
 
 	@ClassRule
@@ -52,19 +57,41 @@ public class PollsQuestionStagedModelDataHandlerTest
 			TransactionalTestRule.INSTANCE);
 
 	@Override
+	protected Map<String, List<StagedModel>> addDependentStagedModelsMap(
+			Group group)
+		throws Exception {
+
+		Map<String, List<StagedModel>> dependentStagedModelsMap =
+			new HashMap<>();
+
+		PollsQuestion question = PollsTestUtil.addQuestion(group.getGroupId());
+
+		addDependentStagedModel(
+			dependentStagedModelsMap, PollsQuestion.class, question);
+
+		return dependentStagedModelsMap;
+	}
+
+	@Override
 	protected StagedModel addStagedModel(
 			Group group,
 			Map<String, List<StagedModel>> dependentStagedModelsMap)
 		throws Exception {
 
-		return PollsTestUtil.addQuestion(group.getGroupId());
+		List<StagedModel> dependentStagedModels = dependentStagedModelsMap.get(
+			PollsQuestion.class.getSimpleName());
+
+		PollsQuestion question = (PollsQuestion)dependentStagedModels.get(0);
+
+		return PollsTestUtil.addChoice(
+			group.getGroupId(), question.getQuestionId());
 	}
 
 	@Override
 	protected StagedModel getStagedModel(String uuid, Group group) {
 		try {
-			return PollsQuestionLocalServiceUtil.
-				getPollsQuestionByUuidAndGroupId(uuid, group.getGroupId());
+			return PollsChoiceLocalServiceUtil.getPollsChoiceByUuidAndGroupId(
+				uuid, group.getGroupId());
 		}
 		catch (Exception e) {
 			return null;
@@ -73,7 +100,24 @@ public class PollsQuestionStagedModelDataHandlerTest
 
 	@Override
 	protected Class<? extends StagedModel> getStagedModelClass() {
-		return PollsQuestion.class;
+		return PollsChoice.class;
+	}
+
+	@Override
+	protected void validateImport(
+			Map<String, List<StagedModel>> dependentStagedModelsMap,
+			Group group)
+		throws Exception {
+
+		List<StagedModel> dependentStagedModels = dependentStagedModelsMap.get(
+			PollsQuestion.class.getSimpleName());
+
+		Assert.assertEquals(1, dependentStagedModels.size());
+
+		PollsQuestion question = (PollsQuestion)dependentStagedModels.get(0);
+
+		PollsQuestionLocalServiceUtil.getPollsQuestionByUuidAndGroupId(
+			question.getUuid(), group.getGroupId());
 	}
 
 	@Override
@@ -83,16 +127,13 @@ public class PollsQuestionStagedModelDataHandlerTest
 
 		super.validateImportedStagedModel(stagedModel, importedStagedModel);
 
-		PollsQuestion question = (PollsQuestion)stagedModel;
-		PollsQuestion importedQuestion = (PollsQuestion)importedStagedModel;
+		PollsChoice choice = (PollsChoice)stagedModel;
+		PollsChoice importedChoice = (PollsChoice)importedStagedModel;
 
-		Assert.assertEquals(question.getTitle(), importedQuestion.getTitle());
+		Assert.assertEquals(choice.getName(), importedChoice.getName());
 		Assert.assertEquals(
-			question.getDescription(), importedQuestion.getDescription());
-		Assert.assertEquals(
-			question.getExpirationDate(), importedQuestion.getExpirationDate());
-		Assert.assertEquals(
-			question.getLastVoteDate(), importedQuestion.getLastVoteDate());
+			choice.getDescription(LocaleUtil.getDefault()),
+			importedChoice.getDescription(LocaleUtil.getDefault()));
 	}
 
 }
