+++ modules/apps/forms-and-workflow/polls/polls-test/src/testIntegration/java/com/liferay/polls/internal/exportimport/data/handler/test/PollsPortletDataHandlerTest.java
@@ -12,7 +12,7 @@
  * details.
  */
 
-package com.liferay.polls.lar.test;
+package com.liferay.polls.internal.exportimport.data.handler.test;
 
 import com.liferay.arquillian.extension.junit.bridge.junit.Arquillian;
 import com.liferay.polls.constants.PollsPortletKeys;
