+++ modules/apps/forms-and-workflow/polls/polls-test/src/testIntegration/java/com/liferay/polls/internal/exportimport/data/handler/test/PollsVoteStagedModelDataHandlerTest.java
@@ -12,13 +12,16 @@
  * details.
  */
 
-package com.liferay.polls.lar.test;
+package com.liferay.polls.internal.exportimport.data.handler.test;
 
 import com.liferay.arquillian.extension.junit.bridge.junit.Arquillian;
 import com.liferay.polls.model.PollsChoice;
 import com.liferay.polls.model.PollsQuestion;
+import com.liferay.polls.model.PollsVote;
 import com.liferay.polls.service.PollsChoiceLocalServiceUtil;
 import com.liferay.polls.service.PollsQuestionLocalServiceUtil;
+import com.liferay.polls.service.PollsVoteLocalServiceUtil;
+import com.liferay.polls.service.persistence.PollsChoiceUtil;
 import com.liferay.polls.util.test.PollsTestUtil;
 import com.liferay.portal.kernel.model.Group;
 import com.liferay.portal.kernel.model.StagedModel;
@@ -26,7 +29,6 @@ import com.liferay.portal.kernel.test.rule.AggregateTestRule;
 import com.liferay.portal.kernel.test.rule.Sync;
 import com.liferay.portal.kernel.test.rule.SynchronousDestinationTestRule;
 import com.liferay.portal.kernel.test.rule.TransactionalTestRule;
-import com.liferay.portal.kernel.util.LocaleUtil;
 import com.liferay.portal.lar.test.BaseStagedModelDataHandlerTestCase;
 import com.liferay.portal.test.rule.LiferayIntegrationTestRule;
 
@@ -45,7 +47,7 @@ import org.junit.runner.RunWith;
  */
 @RunWith(Arquillian.class)
 @Sync
-public class PollsChoiceStagedModelDataHandlerTest
+public class PollsVoteStagedModelDataHandlerTest
 	extends BaseStagedModelDataHandlerTestCase {
 
 	@ClassRule
@@ -69,6 +71,14 @@ public class PollsChoiceStagedModelDataHandlerTest
 		addDependentStagedModel(
 			dependentStagedModelsMap, PollsQuestion.class, question);
 
+		PollsChoice choice = PollsTestUtil.addChoice(
+			group.getGroupId(), question.getQuestionId());
+
+		PollsChoiceUtil.update(choice);
+
+		addDependentStagedModel(
+			dependentStagedModelsMap, PollsChoice.class, choice);
+
 		return dependentStagedModelsMap;
 	}
 
@@ -78,19 +88,25 @@ public class PollsChoiceStagedModelDataHandlerTest
 			Map<String, List<StagedModel>> dependentStagedModelsMap)
 		throws Exception {
 
-		List<StagedModel> dependentStagedModels = dependentStagedModelsMap.get(
-			PollsQuestion.class.getSimpleName());
+		List<StagedModel> questionDependentStagedModels =
+			dependentStagedModelsMap.get(PollsQuestion.class.getSimpleName());
 
-		PollsQuestion question = (PollsQuestion)dependentStagedModels.get(0);
+		PollsQuestion question =
+			(PollsQuestion)questionDependentStagedModels.get(0);
 
-		return PollsTestUtil.addChoice(
-			group.getGroupId(), question.getQuestionId());
+		List<StagedModel> choiceDependentStagedModels =
+			dependentStagedModelsMap.get(PollsChoice.class.getSimpleName());
+
+		PollsChoice choice = (PollsChoice)choiceDependentStagedModels.get(0);
+
+		return PollsTestUtil.addVote(
+			group.getGroupId(), question.getQuestionId(), choice.getChoiceId());
 	}
 
 	@Override
 	protected StagedModel getStagedModel(String uuid, Group group) {
 		try {
-			return PollsChoiceLocalServiceUtil.getPollsChoiceByUuidAndGroupId(
+			return PollsVoteLocalServiceUtil.getPollsVoteByUuidAndGroupId(
 				uuid, group.getGroupId());
 		}
 		catch (Exception e) {
@@ -100,7 +116,7 @@ public class PollsChoiceStagedModelDataHandlerTest
 
 	@Override
 	protected Class<? extends StagedModel> getStagedModelClass() {
-		return PollsChoice.class;
+		return PollsVote.class;
 	}
 
 	@Override
@@ -109,31 +125,26 @@ public class PollsChoiceStagedModelDataHandlerTest
 			Group group)
 		throws Exception {
 
-		List<StagedModel> dependentStagedModels = dependentStagedModelsMap.get(
-			PollsQuestion.class.getSimpleName());
+		List<StagedModel> choiceDependentStagedModels =
+			dependentStagedModelsMap.get(PollsChoice.class.getSimpleName());
 
-		Assert.assertEquals(1, dependentStagedModels.size());
+		Assert.assertEquals(1, choiceDependentStagedModels.size());
 
-		PollsQuestion question = (PollsQuestion)dependentStagedModels.get(0);
+		PollsChoice choice = (PollsChoice)choiceDependentStagedModels.get(0);
+
+		PollsChoiceLocalServiceUtil.getPollsChoiceByUuidAndGroupId(
+			choice.getUuid(), group.getGroupId());
+
+		List<StagedModel> questionDependentStagedModels =
+			dependentStagedModelsMap.get(PollsQuestion.class.getSimpleName());
+
+		Assert.assertEquals(1, questionDependentStagedModels.size());
+
+		PollsQuestion question =
+			(PollsQuestion)questionDependentStagedModels.get(0);
 
 		PollsQuestionLocalServiceUtil.getPollsQuestionByUuidAndGroupId(
 			question.getUuid(), group.getGroupId());
-	}
-
-	@Override
-	protected void validateImportedStagedModel(
-			StagedModel stagedModel, StagedModel importedStagedModel)
-		throws Exception {
-
-		super.validateImportedStagedModel(stagedModel, importedStagedModel);
-
-		PollsChoice choice = (PollsChoice)stagedModel;
-		PollsChoice importedChoice = (PollsChoice)importedStagedModel;
-
-		Assert.assertEquals(choice.getName(), importedChoice.getName());
-		Assert.assertEquals(
-			choice.getDescription(LocaleUtil.getDefault()),
-			importedChoice.getDescription(LocaleUtil.getDefault()));
 	}
 
 }
